Repo Housekeeping
The .gitattributes file should not be in the repository. It sets
representation of files on the reader's system, and they are not the
business of the code author's to set. Also, this wreaks havok with the
checked in png, ttf and sketch files. A properly set up system does not
specify any setting for png files, meaning git should not try any detection
or conversion. The system should have a system-level setting for *.cs
files. If it is unix-ish system, its system setting should be "text", which
implies do whatever EOL conversion is appropriate on *that system* which
in this case would mean "none". If the system is window-ish, the same
system-setting of "text" would mean convert to CRLF in the local tree,
and back to LF in the git repo. The repo therefore has no business setting
this, and in any case the only not-wrong setting would be "*.cs text".
nothing more.