Packaging for Harmony-CitiesSkylines (v2 lib)
1. Enable Strong Names
This will be used to detect which DLLs are shipped with the mod, and
prefer them to third party assemblies

2. Bring MonoMod.Common local so it can be Signed.
Since Harmony-v2 is signed, its dependency MonoMod.Commong must be signed
because it needs Internals visibility.

The nuget package for MonoMod.Common is unsigned, and is not useful.

Hence, to enable StrongNames in this dependency it must be a submodule

Packing is also disabled because it causes the packed Assembly to not be
StrongNamed. I don't know why.

3. Enable Debug Symbols in all but Release
This enables Beta to have symbols

4. The project rename is necessary so Harmony 2 and Harmony 1 can coexist
with in the same solution