Add enabling of the Harmony API
In order to block mods from using the Harmony Lib directly if the Harmony
Mod is disabled, the lib must be disabled by default, and enabled when
the Mod is enabled.

This enables the mod to verify that there are no conficts between other
mods and the Harmony Mod or this library fork. If there are, the player
is informed with actionable diagnostics how to fix the conflict.

Without this gate, players might disable the Harmony Mod and the Lib would
still work with their other mods, but not in a provable, traceable manner.
This lock enables that provability.

The predecessor mod from boformer would work equally whether enabled or
disabled, and the community expects that the mod need not be enabled for
it to work. The lock will also change this mistaken expectation.

The Harmony Lib will requests Awareness Status on first use. This gives
the HarmonyMod the chance to patch Harmony V1 instances, and possibly
disable the Harmony Lib if needed (eg, if there are conflicting
assemblies loaded). The mod need not be loaded first.

Two static vars are stored in the HarmonyLib, acting as a registry for
the several possible IAmAware instances: unsupportedException and
Harmony1Patched, as there is a single current Harmony lib that all
IAmAware's reference