Implement ACL (2.0.1.0 backport)
Access Control List is needed to prevent user mods from patching or
(accidentally) removing Harmony patches. Some mods use UnpatchAll() as
a quick and easy way to clean up when OnDisabled(). This removes patches
that all mods have installed, and also removes Harmony Mods' own patches
to Harmony v1.x and to the game, which breaks the Harmony Mod itself, as
well as the other mods that did not expect to have their own patches
disappear.

This commit causes HarmonyLib to make an ACL request for verification to
the Awareness interface before installing or removing a patch.
The Awareness call should throw if the patch should not proceed.

# Conflicts:
#	Harmony/Public/PatchProcessor.cs